try to make Output an abstract class
