+++ Output.py
@@ -1,60 +1,73 @@
-from Item import Weapon
+from abc import ABC, abstractmethod
 from Player import Player
 from Data import *
 from time import sleep
-import telepot
-from telepot import message_identifier
-from telepot.exception import TelegramError, TooManyRequestsError, BotWasBlockedError
-from telepot.namedtuple import InlineKeyboardMarkup, InlineKeyboardButton
 
-helpString = """join - Join a game
-start - Start a game
-jizz - Throw a die to move
-map - Show the game map
-pos - Show Your positon
-buy - [no] Buy an item in shop
-upgrade - [weapon/armor] Upgrade weapon or armor in blacksmith shop
-change - Change weapon or armor
-end - Leave shop or blacksmith shop
-drink - [no] Drink a potion
-mystat - Show your status
-showpotion - Show all your potions
-showstat - Show status of chosen entity
-retire - retire from game
-help - Show game help"""
-
-def sending(func):
-    def sending_wrapper(*args, **kwargs):
-        try:
-            return func(*args, **kwargs)
-        except TooManyRequestsError:
-            sleep(5)
-            return sending_wrapper(*args, **kwargs)
-        except (BotWasBlockedError, TelegramError):
-            #pm failed
-            #args[0] is self
-            args[0].bot.sendMessage(args[0].id, "Please add me! @inforJourneyBot")
-    return sending_wrapper
-
-class Output:
-    def __init__(self, bot: telepot.Bot, id):
+class Output(ABC):
+    @staticmethod
+    @abstractmethod
+    def BadRequest() -> type:
+        pass
+    
+    def __init__(self, bot, id, decorator):
         self.bot = bot
         self.id = id
-        self.end_markup = InlineKeyboardMarkup(inline_keyboard=[
-            [InlineKeyboardButton(text='Leave', callback_data='end')],
+        self.end_markup = self.gen_inline_keyboard_barkup(inline_keyboard=[
+            [self.gen_inline_keyboard_button(text='Leave', callback_data='end')],
         ])
+        print("calling Output.__init__")
+        for name in dir(self):
+            if name.startswith("send") or name.startswith("stat"):
+                setattr(self, name, decorator(getattr(self, name)))
+
+    @abstractmethod
+    def _send_message():
+        pass
+
+    @abstractmethod
+    def _edit_message_text():
+        pass
+
+    @staticmethod
+    @abstractmethod
+    def gen_inline_keyboard_barkup():
+        pass
     
+    @staticmethod
+    @abstractmethod
+    def gen_inline_keyboard_button():
+        pass
+
+    @abstractmethod
+    def _send_photo():
+        pass
+
     @sending
     def send_help(self):
-        self.bot.sendMessage(self.id, helpString)
+        helpString = "join - Join a game\n" + \
+            "start - Start a game\n" + \
+            "jizz - Throw a die to move\n" + \
+            "map - Show the game map\n" + \
+            "pos - Show Your positon\n" + \
+            "buy - [no] Buy an item in shop\n" + \
+            "upgrade - [weapon/armor] Upgrade weapon or armor in blacksmith shop\n" + \
+            "change - Change weapon or armor\n" + \
+            "end - Leave shop or blacksmith shop\n" + \
+            "drink - [no] Drink a potion\n" + \
+            "mystat - Show your status\n" + \
+            "showpotion - Show all your potions\n" + \
+            "showstat - Show status of chosen entity\n" + \
+            "retire - retire from game\n" + \
+            "help - Show game help\n"
+        self._send_message(self.id, helpString)
 
     @sending
     def send_welcome(self, name):
-        self.bot.sendMessage(self.id, f"{name}加入了遊戲", parse_mode="Markdown")
+        self._send_message(self.id, f"{name}加入了遊戲", parse_mode="Markdown")
 
     @sending
     def send_start_game(self):
-        self.bot.sendMessage(self.id, "遊戲已開始")
+        self._send_message(self.id, "遊戲已開始")
 
     def tag_user(self, player):
         if player.username:
@@ -64,22 +77,22 @@ class Output:
 
     @sending
     def send_player_turn_start(self, player: Player):
-        self.bot.sendMessage(self.id, 
+        self._send_message(self.id, 
             f"換{player.name}了唷 {self.tag_user(player)}\n你目前在第{player.pos}格", 
             parse_mode="Markdown"
         )
     
     @sending
     def send_jizz_result(self, player_name, jizz):
-        self.bot.sendMessage(self.id, f"{player_name}骰出了{jizz}.")
+        self._send_message(self.id, f"{player_name}骰出了{jizz}.")
     
     @sending
     def send_map(self):
-        self.bot.sendPhoto(self.id, open("./Img/raw_map.jpg","rb"))
+        self._send_photo(self.id, open("./Img/raw_map.jpg","rb"))
     
     @sending
     def send_pos(self, name, pos):
-        self.bot.sendMessage(self.id, f"{name}目前在第{pos}格")
+        self._send_message(self.id, f"{name}目前在第{pos}格")
 
     @sending
     def send_potion(self, uid, name, potions):
@@ -89,141 +102,158 @@ class Output:
         else:
             potions_list = [f"{i}. {str(potion)}" for i, potion in enumerate(potions)]
             potions_str = "\n".join(potions_list)
-        self.bot.sendMessage(uid, f"{name}的藥水們:\n{potions_str}")
+        self._send_message(uid, f"{name}的藥水們:\n{potions_str}")
 
     @sending
     def send_heal_result(self, player: Player, potion, heal_point):
-        self.bot.sendMessage(self.id, f"{player.name}飲用了{potion}\n回復{heal_point}點生命\n{player.name}現在有{player.hp}點生命")
+        self._send_message(self.id, f"{player.name}飲用了{potion}\n回復{heal_point}點生命\n{player.name}現在有{player.hp}點生命")
 
-    @sending 
+    @sending
     def send_wrong_argument(self):
-        self.bot.sendMessage(self.id, "參數錯誤")
+        self._send_message(self.id, "參數錯誤")
 
     @sending
     def send_change(self, name, uid, avaliable_items):
         kb_list = []
         for w in avaliable_items:
-            kb_list.append([InlineKeyboardButton(text=w.name, callback_data="change "+str(uid)+" "+w.name)])
+            kb_list.append([self.gen_inline_keyboard_button(text=w.name, callback_data="change "+str(uid)+" "+w.name)])
 
         if kb_list:
-            keyboard = InlineKeyboardMarkup(inline_keyboard=kb_list)
-            self.bot.sendMessage(self.id, f"{name}, 請選擇更換裝備", reply_markup=keyboard)
+            keyboard = self.gen_inline_keyboard_markup(inline_keyboard=kb_list)
+            self._send_message(self.id, f"{name}, 請選擇更換裝備", reply_markup=keyboard)
         else:
-            self.bot.sendMessage(self.id, f"{name}沒有可更換的裝備")
+            self._send_message(self.id, f"{name}沒有可更換的裝備")
 
-    @sending 
+    @sending
     def change_succeed(self, name, item, identifier):
-        self.bot.editMessageText(identifier, f"{name}已成功裝備{item}")
-       
+        self._edit_message_text(f"{name}已成功裝備{item}", identifier=identifier)
 
     @sending
     def send_retire_confirm(self, name):
-        self.bot.sendMessage(self.id, f"{name},你確定要退出遊戲嗎?\n確定退出請再次輸入 /retire")
+        self._send_message(self.id, f"{name},你確定要退出遊戲嗎?\n確定退出請再次輸入 /retire")
 
     @sending
     def send_retire(self, name):
-        self.bot.sendMessage(self.id, f"{name}離開了遊戲")
+        self._send_message(self.id, f"{name}離開了遊戲")
     
     @sending
     def send_stat(self, uid):
         kb_list = [
-            [InlineKeyboardButton(text="小怪", callback_data="showstat monster"),
-            InlineKeyboardButton(text="首領", callback_data="showstat boss")],
-            [InlineKeyboardButton(text="玩家", callback_data="showstat player"),
-            InlineKeyboardButton(text="武器", callback_data="showstat weapon 0")],
-            [InlineKeyboardButton(text="防具", callback_data="showstat armor 0"),
-            InlineKeyboardButton(text="道具", callback_data="showstat item")]
+            [self.gen_inline_keyboard_button(text="小怪", callback_data="showstat monster"),
+            self.gen_inline_keyboard_button(text="首領", callback_data="showstat boss")],
+            [self.gen_inline_keyboard_button(text="玩家", callback_data="showstat player"),
+            self.gen_inline_keyboard_button(text="武器", callback_data="showstat weapon 0")],
+            [self.gen_inline_keyboard_button(text="防具", callback_data="showstat armor 0"),
+            self.gen_inline_keyboard_button(text="道具", callback_data="showstat item")]
         ]
-        keyboard = InlineKeyboardMarkup(inline_keyboard=kb_list)
-        self.bot.sendMessage(uid, "請選擇種類", reply_markup=keyboard)
+        keyboard = self.gen_inline_keyboard_markup(inline_keyboard=kb_list)
+        self._send_message(uid, "請選擇種類", reply_markup=keyboard)
 
     @sending
     def stat_category(self, identifier):
         kb_list = [
-            [InlineKeyboardButton(text="小怪", callback_data="showstat monster"),
-            InlineKeyboardButton(text="首領", callback_data="showstat boss")],
-            [InlineKeyboardButton(text="玩家", callback_data="showstat player"),
-            InlineKeyboardButton(text="武器", callback_data="showstat weapon 0")],
-            [InlineKeyboardButton(text="防具", callback_data="showstat armor 0"),
-            InlineKeyboardButton(text="道具", callback_data="showstat item")]
+            [self.gen_inline_keyboard_button(text="小怪", callback_data="showstat monster"),
+            self.gen_inline_keyboard_button(text="首領", callback_data="showstat boss")],
+            [self.gen_inline_keyboard_button(text="玩家", callback_data="showstat player"),
+            self.gen_inline_keyboard_button(text="武器", callback_data="showstat weapon 0")],
+            [self.gen_inline_keyboard_button(text="防具", callback_data="showstat armor 0"),
+            self.gen_inline_keyboard_button(text="道具", callback_data="showstat item")]
         ]
-        keyboard = InlineKeyboardMarkup(inline_keyboard=kb_list)
-        self.bot.editMessageText(identifier, "請選擇種類", reply_markup=keyboard)
+        keyboard = self.gen_inline_keyboard_markup(inline_keyboard=kb_list)
+        self._edit_message_text("請選擇種類", identifier=identifier, reply_markup=keyboard)
 
     @sending
     def stat_monster_stage(self, identifier):
         kb_list = []
         for i in range(4):
-            kb_list.append([InlineKeyboardButton(text="階段"+str(i+1), callback_data="showstat monster "+str(i))])
-        kb_list.append([InlineKeyboardButton(text="上一層", callback_data="showstat")])
-        keyboard = InlineKeyboardMarkup(inline_keyboard=kb_list)
-        self.bot.editMessageText(identifier, "請選擇", reply_markup=keyboard)
+            kb_list.append([self.gen_inline_keyboard_button(text=f"階段{i+1}", callback_data=f"showstat monster {i}")])
+        kb_list.append([self.gen_inline_keyboard_button(text="上一層", callback_data="showstat")])
+        keyboard = self.gen_inline_keyboard_markup(inline_keyboard=kb_list)
+        self._edit_message_text("請選擇", identifier=identifier, reply_markup=keyboard)
 
     @sending
     def stat_monsters(self, identifier, stage):
         kb_list = []
-        for name in Monsters[stage]:
-            kb_list.append([InlineKeyboardButton(text=name, callback_data="showstat monster "+stage+" "+name)])
-        kb_list.append([InlineKeyboardButton(text="上一層", callback_data="showstat monster")])
-        keyboard = InlineKeyboardMarkup(inline_keyboard=kb_list)
-        self.bot.editMessageText(identifier, "請選擇", reply_markup=keyboard)
+        display = list(Monsters[stage])
+        for name1, name2 in zip(display[::2], display[1::2]):
+            kb_list.append([
+                self.gen_inline_keyboard_button(text=name1, callback_data=f"showstat monster {stage} {name1}"), 
+                self.gen_inline_keyboard_button(text=name2, callback_data=f"showstat monster {stage} {name2}")
+            ])
+        if len(display) % 2:
+            kb_list.append([self.gen_inline_keyboard_button(text=display[-1], callback_data=f"showstat monster {stage} {display[-1]}")])
+        kb_list.append([self.gen_inline_keyboard_button(text="上一層", callback_data="showstat monster")])
+        keyboard = self.gen_inline_keyboard_markup(inline_keyboard=kb_list)
+        self._edit_message_text("請選擇", identifier=identifier, reply_markup=keyboard)
 
     @sending
     def stat_bosses(self, identifier):
         kb_list = []
         for i in range(4):
-            kb_list.append([InlineKeyboardButton(text=Bosses[i][0], callback_data="showstat boss "+str(i))])
-        kb_list.append([InlineKeyboardButton(text="上一層", callback_data="showstat")])
-        keyboard = InlineKeyboardMarkup(inline_keyboard=kb_list)
-        self.bot.editMessageText(identifier, "請選擇", reply_markup=keyboard)
+            kb_list.append([self.gen_inline_keyboard_button(text=Bosses[i][0], callback_data=f"showstat boss {i}")])
+        kb_list.append([self.gen_inline_keyboard_button(text="上一層", callback_data="showstat")])
+        keyboard = self.gen_inline_keyboard_markup(inline_keyboard=kb_list)
+        self._edit_message_text("請選擇", identifier=identifier, reply_markup=keyboard)
 
     @sending
     def stat_players(self, players, identifier):
         kb_list = []
         for player in players:
-            kb_list.append([InlineKeyboardButton(text=player.name, callback_data="showplayer "+str(player.id))])
-        kb_list.append([InlineKeyboardButton(text="上一層", callback_data="showstat")])
-        keyboard = InlineKeyboardMarkup(inline_keyboard=kb_list)
-        self.bot.editMessageText(identifier, "請選擇", reply_markup=keyboard)
+            kb_list.append([self.gen_inline_keyboard_button(text=player.name, callback_data=f"showplayer {player.id}")])
+        kb_list.append([self.gen_inline_keyboard_button(text="上一層", callback_data="showstat")])
+        keyboard = self.gen_inline_keyboard_markup(inline_keyboard=kb_list)
+        self._edit_message_text("請選擇", identifier=identifier, reply_markup=keyboard)
 
     @sending
     def stat_items(self, identifier):
         kb_list = []
         for p in Potions:
-            kb_list.append([InlineKeyboardButton(text=p, callback_data="showstat item "+p)])
-        kb_list.append([InlineKeyboardButton(text="上一層", callback_data="showstat")])
-        keyboard = InlineKeyboardMarkup(inline_keyboard=kb_list)
-        self.bot.editMessageText(identifier, "請選擇", reply_markup=keyboard)
+            kb_list.append([self.gen_inline_keyboard_button(text=p, callback_data=f"showstat item {p}")])
+        kb_list.append([self.gen_inline_keyboard_button(text="上一層", callback_data="showstat")])
+        keyboard = self.gen_inline_keyboard_markup(inline_keyboard=kb_list)
+        self._edit_message_text("請選擇", identifier=identifier, reply_markup=keyboard)
 
     @sending
     def stat_weapons(self, identifier, page):
         kb_list = []
         weapon_list = list(Weapons.items())
         if page:
-            kb_list.append([InlineKeyboardButton(text="<<", callback_data="showstat weapon "+str(page-1))])
-        for i in range(page*6, min(page*6+6, len(weapon_list))):
-            name, param = weapon_list[i]
-            kb_list.append([InlineKeyboardButton(text=name, callback_data="showweapon "+name)])
+            kb_list.append([self.gen_inline_keyboard_button(text="<<", callback_data=f"showstat weapon {page-1}")])
+
+        display = weapon_list[page*6:min(page*6+6, len(weapon_list))]
+        for item1, item2 in zip(display[::2], display[1::2]):
+            kb_list.append([
+                self.gen_inline_keyboard_button(text=item1[0], callback_data=f"showweapon {item1[0]}"),
+                self.gen_inline_keyboard_button(text=item2[0], callback_data=f"showweapon {item2[0]}")
+            ])
+        if len(display) % 2:
+            kb_list.append([self.gen_inline_keyboard_button(text=display[-1][0], callback_data=f"showweapon {display[-1][0]}")])
         if page*6 + 6 < len(weapon_list):
-            kb_list.append([InlineKeyboardButton(text=">>", callback_data="showstat weapon "+str(page+1))])
-        kb_list.append([InlineKeyboardButton(text="上一層", callback_data="showstat")])
-        keyboard = InlineKeyboardMarkup(inline_keyboard=kb_list)
-        self.bot.editMessageText(identifier, "請選擇", reply_markup=keyboard)
+            kb_list.append([self.gen_inline_keyboard_button(text=">>", callback_data=f"showstat weapon {page+1}")])
+        kb_list.append([self.gen_inline_keyboard_button(text="上一層", callback_data="showstat")])
+        keyboard = self.gen_inline_keyboard_markup(inline_keyboard=kb_list)
+        self._edit_message_text("請選擇", identifier=identifier, reply_markup=keyboard)
 
     @sending
     def stat_armors(self, identifier, page):
         kb_list = []
         armor_list = list(Armors.items())
         if page:
-            kb_list.append([InlineKeyboardButton(text="<<", callback_data="showstat armor "+str(page-1))])
-        for i in range(page*6, min(page*6+6, len(armor_list))):
-            name, param = armor_list[i]
-            kb_list.append([InlineKeyboardButton(text=name, callback_data="showarmor "+name)])
+            kb_list.append([self.gen_inline_keyboard_button(text="<<", callback_data=f"showstat armor {page-1}")])
+
+        display = armor_list[page*6:min(page*6+6, len(armor_list))]
+        for item1, item2 in zip(display[::2], display[1::2]):
+            kb_list.append([
+                self.gen_inline_keyboard_button(text=item1[0], callback_data=f"showarmor {item1[0]}"),
+                self.gen_inline_keyboard_button(text=item2[0], callback_data=f"showarmor {item2[0]}")
+            ])
+        if len(display) % 2:
+            kb_list.append([self.gen_inline_keyboard_button(text=display[-1][0], callback_data=f"showarmor {display[-1][0]}")])
         if page*6 + 6 < len(armor_list):
-            kb_list.append([InlineKeyboardButton(text=">>", callback_data="showstat armor "+str(page+1))])
-        kb_list.append([InlineKeyboardButton(text="上一層", callback_data="showstat")])
-        keyboard = InlineKeyboardMarkup(inline_keyboard=kb_list)
-        self.bot.editMessageText(identifier, "請選擇", reply_markup=keyboard)
+            kb_list.append([self.gen_inline_keyboard_button(text=">>", callback_data=f"showstat armor {page+1}")])
+        kb_list.append([self.gen_inline_keyboard_button(text="上一層", callback_data="showstat")])
+        keyboard = self.gen_inline_keyboard_markup(inline_keyboard=kb_list)
+        self._edit_message_text("請選擇", identifier=identifier, reply_markup=keyboard)
 
     @sending
     def stat_player(self, uid, player: Player):
@@ -247,11 +277,11 @@ class Output:
             f'{player.name}現在有 {player.coin} 金幣'
         )
 
-        self.bot.sendMessage(uid, player_str)
+        self._send_message(uid, player_str)
     
     @sending
     def stat_monster(self, uid, monster_name, monster_data):
-        self.bot.sendMessage(uid, (
+        self._send_message(uid, (
             f'{monster_name}: 攻:{monster_data[0]}, 防:{monster_data[1]}, HP: {monster_data[2]}\n'
             f'經驗值: {monster_data[3]}, 金幣: {monster_data[4]}\n'
             f'出現等級: {monster_data[5]} ~ {monster_data[6]}'
@@ -259,32 +289,37 @@ class Output:
 
     @sending
     def stat_item(self, uid, potion):
-        self.bot.sendMessage(uid, f'恢復{"/".join(map(lambda x: str(x), potion))}點生命')
+        self._send_message(uid, f'恢復{"/".join(map(lambda x: str(x), potion))}點生命')
 
     @sending
     def stat_weapon(self, uid, weapon):
         (atk, dfd) = Weapons[weapon]
-        self.bot.sendMessage(uid, f"{weapon}:\n攻:{atk} 防:{dfd}")
+        self._send_message(uid, f"{weapon}:\n攻:{atk} 防:{dfd}")
 
     @sending
     def stat_armor(self, uid, armor):
         (atk, dfd) = Armors[armor]
-        self.bot.sendMessage(uid, f"{armor}:\n攻:{atk} 防:{dfd}")
+        self._send_message(uid, f"{armor}:\n攻:{atk} 防:{dfd}")
 
     @sending
     def send_upgrade_limited(self, upgrade_times):
-        self.bot.sendMessage(self.id, f"金幣不足 共升級{upgrade_times}次")
+        self._send_message(self.id, f"金幣不足 共升級{upgrade_times}次")
 
     @sending
     def send_upgrade_full(self, upgrade_times):
-        self.bot.sendMessage(self.id, f"升級{upgrade_times}次成功")
+        self._send_message(self.id, f"升級{upgrade_times}次成功")
 
+    @abstractmethod
     def _editable_send(self, content, message):
-        if message == None:
-            print("message is None")
-        if not message:
-            return self.bot.sendMessage(self.id, content)
-        return self.bot.editMessageText(message_identifier(message), f"{message['text']}\n{content}")
+        try:
+            if message == None:
+                print("message is None")
+            if not message:
+                return self._send_message(self.id, content)
+            return self._edit_message_text(f"{message.text}\n{content}", message=message)
+        except self.BadRequest():
+            print("Bad request error")
+            return message
 
     @sending
     def send_meet(self, name, enemy, event_name, message = None):
@@ -336,8 +371,8 @@ class Output:
             msg+= "{}. {}{:>5}\n".format(i,msgstr,price)
         msg+="```"
         if not message:
-            return self.bot.sendMessage(self.id, msg, parse_mode="Markdown", reply_markup=self.end_markup)
-        return self.bot.editMessageText(message_identifier(message), f"{message['text']}\n{msg}", parse_mode="Markdown", reply_markup=self.end_markup)
+            return self._send_message(self.id, msg, parse_mode="Markdown", reply_markup=self.end_markup)
+        return self._edit_message_text(f"{message.text}\n{msg}", message=message, parse_mode="Markdown", reply_markup=self.end_markup)
 
     @sending
     def send_reach_blacksmith(self, name, message = None):
@@ -350,8 +385,8 @@ class Output:
                f"防具升級 攻防+{upgrade} 價格:{armor_cost}金幣\n"
         )
         if not message:
-            return self.bot.sendMessage(self.id, content, reply_markup=self.end_markup)
-        return self.bot.editMessageText(message_identifier(message), f"{message['text']}\n{content}", reply_markup=self.end_markup)
+            return self._send_message(self.id, content, reply_markup=self.end_markup)
+        return self._edit_message_text(f"{message.text}\n{content}", message=message, parse_mode="Markdown", reply_markup=self.end_markup)
 
     @sending
     def send_find_chest(self, name, coin, weapon_name = None, message = None):
@@ -366,12 +401,69 @@ class Output:
 
     @sending
     def send_not_enough_coin(self):
-        self.bot.sendMessage(self.id, "金幣不足 購買失敗")
+        self._send_message(self.id, "金幣不足 購買失敗")
 
     @sending
     def send_buy_success(self):
-        self.bot.sendMessage(self.id, "購買成功")
+        self._send_message(self.id, "購買成功")
 
     @sending
     def send_end_game(self):
-        self.bot.sendMessage(self.id, "遊戲已結束")
+        self._send_message(self.id, "遊戲已結束")
+
+    def direct_from_in(self, game, query_data, uid, identifier):
+        try:
+            if query_data[0] == "change" and len(query_data) > 2:
+                if uid == int(query_data[1]):
+                    game.request_change(uid, query_data[2], identifier)
+                else:
+                    print(uid, query_data)
+            elif query_data[0] == "end":
+                game.request_end(uid)
+            elif query_data[0] == "showstat":
+                try:
+                    if len(query_data) > 3:
+                        if query_data[1] == "monster":
+                            game.show_monster(uid, query_data[3], Monsters[int(query_data[2])][query_data[3]])
+                    if len(query_data) > 2:
+                        if query_data[1] == "monster":
+                            i = int(query_data[2])
+                            self.stat_monsters(identifier, i)
+                        elif query_data[1] == "weapon":
+                            start = int(query_data[2])
+                            self.stat_weapons(identifier, start)
+                        elif query_data[1] == "armor":
+                            start = int(query_data[2])
+                            self.stat_armors(identifier, start)
+                        elif query_data[1] == "boss":
+                            i = int(query_data[2])
+                            game.show_monster(uid, Bosses[i][0], Bosses[i][1:])
+                        elif query_data[1] == "item":
+                            if query_data[2] in Potions:
+                                potion = Potions[query_data[2]]
+                                self.stat_item(uid, potion)
+                    elif len(query_data) > 1:
+                        if query_data[1] == "player":
+                            self.stat_players(game.get_players(), identifier)
+                        elif query_data[1] == "boss":
+                            self.stat_bosses(identifier)
+                        elif query_data[1] == "item":
+                            self.stat_items(identifier)
+                        elif query_data[1] == "monster":
+                            self.stat_monster_stage(identifier)
+                    else:
+                        self.stat_category(identifier)
+                except:
+                    pass
+            elif query_data[0] == "showplayer" and len(query_data)>1:
+                try:
+                    show_player_id = int(query_data[1])
+                    game.request_show_player(uid, show_player_id)
+                except:
+                    pass
+            elif query_data[0] == "showweapon" and len(query_data)>1:
+                self.stat_weapon(uid, query_data[1])
+            elif query_data[0] == "showarmor" and len(query_data)>1:
+                self.stat_armor(uid, query_data[1])
+        except:
+            pass
