+++ server.py
@@ -2,7 +2,7 @@
 # from os import stat
 # from re import UNICODE
 import telegram
-from OutputMean import Output
+from OutputTLP import OutputTLP as Output
 # from urllib3.poolmanager import PoolKey
 # from Events import EventCanEnd
 import telepot
@@ -40,8 +40,7 @@ def handle(msg):
         return
     gid = msg["chat"]["id"]
     if gid not in games:
-        bot_mean = telegram.Bot(TOKEN)
-        out = Output(bot_mean, gid)
+        out = Output(bot, gid)
         games[gid] = Game(gid, out)
     in_data = msg["text"].replace("@inforJourneyBot", "").split()
     if "username" in msg["from"]:
